feat: CoreData UnitTest 진행하기
